Remove unused duplicate courses
Some of the `Course.AddCourse(...)` calls were commented out because they were typically identical to the lines preceding them. The reason why they were originally added if from the spreadsheet of class data: the spreadsheet will, for each _section_ generate the lines of code needed to create the _course_ and _section_. This meant that courses with multiple sections had code generated to create the course again for each additional section. When the code was copied into the database, it was commented out as a duplicate.
It has now been removed, since it did not serve a purpose.